# target/generated-sources/openapi/src/main/java/MazeGameClientApplication/model/GameDto.java
/*
 * Maze Game API
 * A simple single-player maze game.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package MazeGameClientApplication.model;

import java.util.Objects;
import java.util.Arrays;
import MazeGameClientApplication.model.GameStatusDto;
import MazeGameClientApplication.model.PositionDto;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import java.math.BigDecimal;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import com.fasterxml.jackson.annotation.JsonTypeName;

/**
 * GameDto
 */
@JsonPropertyOrder({
  GameDto.JSON_PROPERTY_GAME_ID,
  GameDto.JSON_PROPERTY_GROUP_NAME,
  GameDto.JSON_PROPERTY_POSITION,
  GameDto.JSON_PROPERTY_STATUS,
  GameDto.JSON_PROPERTY_AVAILABLE_MOVES
})
@JsonTypeName("Game")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2025-06-01T23:38:36.918660300+02:00[Europe/Berlin]", comments = "Generator version: 7.10.0")
public class GameDto {
  public static final String JSON_PROPERTY_GAME_ID = "gameId";
  @jakarta.annotation.Nonnull
  private BigDecimal gameId;

  public static final String JSON_PROPERTY_GROUP_NAME = "groupName";
  @jakarta.annotation.Nullable
  private String groupName;

  public static final String JSON_PROPERTY_POSITION = "position";
  @jakarta.annotation.Nonnull
  private PositionDto position;

  public static final String JSON_PROPERTY_STATUS = "status";
  @jakarta.annotation.Nonnull
  private GameStatusDto status;

  /**
   * Gets or Sets availableMoves
   */
  public enum AvailableMovesEnum {
    LEFT(String.valueOf("left")),
    
    UP(String.valueOf("up")),
    
    RIGHT(String.valueOf("right")),
    
    DOWN(String.valueOf("down"));

    private String value;

    AvailableMovesEnum(String value) {
      this.value = value;
    }

    @JsonValue
    public String getValue() {
      return value;
    }

    @Override
    public String toString() {
      return String.valueOf(value);
    }

    @JsonCreator
    public static AvailableMovesEnum fromValue(String value) {
      for (AvailableMovesEnum b : AvailableMovesEnum.values()) {
        if (b.value.equals(value)) {
          return b;
        }
      }
      throw new IllegalArgumentException("Unexpected value '" + value + "'");
    }
  }

  public static final String JSON_PROPERTY_AVAILABLE_MOVES = "availableMoves";
  @jakarta.annotation.Nonnull
  private List<AvailableMovesEnum> availableMoves = new ArrayList<>();

  public GameDto() {
  }

  public GameDto gameId(@jakarta.annotation.Nonnull BigDecimal gameId) {
    
    this.gameId = gameId;
    return this;
  }

  /**
   * Get gameId
   * @return gameId
   */
  @jakarta.annotation.Nonnull
  @JsonProperty(JSON_PROPERTY_GAME_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public BigDecimal getGameId() {
    return gameId;
  }


  @JsonProperty(JSON_PROPERTY_GAME_ID)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setGameId(@jakarta.annotation.Nonnull BigDecimal gameId) {
    this.gameId = gameId;
  }

  public GameDto groupName(@jakarta.annotation.Nullable String groupName) {
    
    this.groupName = groupName;
    return this;
  }

  /**
   * Get groupName
   * @return groupName
   */
  @jakarta.annotation.Nullable
  @JsonProperty(JSON_PROPERTY_GROUP_NAME)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getGroupName() {
    return groupName;
  }


  @JsonProperty(JSON_PROPERTY_GROUP_NAME)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setGroupName(@jakarta.annotation.Nullable String groupName) {
    this.groupName = groupName;
  }

  public GameDto position(@jakarta.annotation.Nonnull PositionDto position) {
    
    this.position = position;
    return this;
  }

  /**
   * Get position
   * @return position
   */
  @jakarta.annotation.Nonnull
  @JsonProperty(JSON_PROPERTY_POSITION)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public PositionDto getPosition() {
    return position;
  }


  @JsonProperty(JSON_PROPERTY_POSITION)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setPosition(@jakarta.annotation.Nonnull PositionDto position) {
    this.position = position;
  }

  public GameDto status(@jakarta.annotation.Nonnull GameStatusDto status) {
    
    this.status = status;
    return this;
  }

  /**
   * Get status
   * @return status
   */
  @jakarta.annotation.Nonnull
  @JsonProperty(JSON_PROPERTY_STATUS)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public GameStatusDto getStatus() {
    return status;
  }


  @JsonProperty(JSON_PROPERTY_STATUS)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setStatus(@jakarta.annotation.Nonnull GameStatusDto status) {
    this.status = status;
  }

  public GameDto availableMoves(@jakarta.annotation.Nonnull List<AvailableMovesEnum> availableMoves) {
    
    this.availableMoves = availableMoves;
    return this;
  }

  public GameDto addAvailableMovesItem(AvailableMovesEnum availableMovesItem) {
    if (this.availableMoves == null) {
      this.availableMoves = new ArrayList<>();
    }
    this.availableMoves.add(availableMovesItem);
    return this;
  }

  /**
   * Get availableMoves
   * @return availableMoves
   */
  @jakarta.annotation.Nonnull
  @JsonProperty(JSON_PROPERTY_AVAILABLE_MOVES)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)

  public List<AvailableMovesEnum> getAvailableMoves() {
    return availableMoves;
  }


  @JsonProperty(JSON_PROPERTY_AVAILABLE_MOVES)
  @JsonInclude(value = JsonInclude.Include.ALWAYS)
  public void setAvailableMoves(@jakarta.annotation.Nonnull List<AvailableMovesEnum> availableMoves) {
    this.availableMoves = availableMoves;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    GameDto game = (GameDto) o;
    return Objects.equals(this.gameId, game.gameId) &&
        Objects.equals(this.groupName, game.groupName) &&
        Objects.equals(this.position, game.position) &&
        Objects.equals(this.status, game.status) &&
        Objects.equals(this.availableMoves, game.availableMoves);
  }

  @Override
  public int hashCode() {
    return Objects.hash(gameId, groupName, position, status, availableMoves);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class GameDto {\n");
    sb.append("    gameId: ").append(toIndentedString(gameId)).append("\n");
    sb.append("    groupName: ").append(toIndentedString(groupName)).append("\n");
    sb.append("    position: ").append(toIndentedString(position)).append("\n");
    sb.append("    status: ").append(toIndentedString(status)).append("\n");
    sb.append("    availableMoves: ").append(toIndentedString(availableMoves)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}